refactor: Move datetime converter to helper file
